src/requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
click==8.1.4
fastapi==0.100.0
flake8==6.0.0
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
pycodestyle==2.10.0
pydantic==2.0.2
pydantic_core==2.1.2
pyflakes==3.0.1
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.18
sqlalchemy-stubs==0.4
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
